moving filter class objects into separate file
