+++ pyfar/dsp/__init__.py
@@ -1,7 +1,4 @@
-from .filter import Filter
-from .filter import FilterFIR
-from .filter import FilterIIR
-from .filter import FilterSOS
+from .classes import (Filter, FilterFIR, FilterIIR, FilterSOS)
 from .dsp import (
     phase, group_delay, spectrogram, wrap_to_2pi, nextpow2)
